% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{adjustment_from_string}
\alias{adjustment_from_string}
\title{Convert character representation of a list of adjustments to list}
\usage{
adjustment_from_string(adjustment_string)
}
\arguments{
\item{adjustment_string}{A character vector representing the list
of adjustments}
}
\value{
A list of character vectors with names of adjustment
    variables
}
\description{
Convert character representation of a list of adjustments to list
}

