% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\docType{data}
\name{adjustment_separator}
\alias{adjustment_separator}
\title{String used in character representation of a list of adjustments}
\format{An object of class \code{character} of length 1.}
\usage{
adjustment_separator
}
\description{
String used in character representation of a list of adjustments
}
\keyword{datasets}

