% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{adjustment_to_string}
\alias{adjustment_to_string}
\title{Convert a list of adjustment variables to a character vector}
\usage{
adjustment_to_string(adjustments)
}
\arguments{
\item{adjustments}{List of character vectors with names of
adjustment variables}
}
\value{
A character vector representing the list of adjustments.
}
\description{
Convert a list of adjustment variables to a character vector
}

