% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{as.character.manyregs_model}
\alias{as.character.manyregs_model}
\title{Coerce a model to type character.}
\usage{
\method{as.character}{manyregs_model}(x, ...)
}
\arguments{
\item{x}{Model object to be coerced.}

\item{\dots}{Ignored.}
}
\value{
Character representation of model.
}
\description{
Coerce a model to type character.
}

