% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{compare_distros}
\alias{compare_distros}
\title{Compare distributions using a Chi-Squared test}
\usage{
compare_distros(column_names, by, data, digits = 4L)
}
\arguments{
\item{column_names}{Column names of variables}

\item{by}{Name of a grouping variable}

\item{data}{Data frame containing variables `column_names` and `by`}

\item{digits}{Number of decimal digits in `pval` column}
}
\value{
A data frame with columns "variable", "by", "pval",
    "pvalue".  The "variable" column contains the name of the
    variables whose distributions are compared between the levels
    of the `by` variable whose name is contained in the "by"
    column.  The "pval" column contains the p-value resulting from
    the comparison using Pearson's Chi-squared test, rounded to
    `digits` decimal digits for better readability.  The "pvalue"
    column contains the exact p-value.
}
\description{
Compare distributions using a Chi-Squared test
}

