% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{compare_means}
\alias{compare_means}
\title{Compare the means of two groups using a t-test}
\usage{
compare_means(column_names, group1, group2, data, digits = 4L)
}
\arguments{
\item{column_names}{Column names of variables}

\item{group1}{Logical vector defining observations in 1st group}

\item{group2}{Logical vector defining observations in 2nd group}

\item{data}{Data frame containing variables `column_names`}

\item{digits}{Number of decimal digits in `pval` column}
}
\value{
A data frame with columns "variable", "pval", "pvalue".
    The "variable" column contains the name of the variable whose
    means were compared between `group1` and `group2`.  The "pval"
    column contains the p-value resulting comparing the means of
    `group1` and `group2` using a t-test, rounded to `digits`
    decimal digits for better readability.  The "pvalue" column
    contains the exact p-value.
}
\description{
Compare the means of two groups using a t-test
}

