% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\docType{data}
\name{covariate_separator}
\alias{covariate_separator}
\title{String separating names of covariates in formula string.}
\format{An object of class \code{character} of length 1.}
\usage{
covariate_separator
}
\description{
String separating names of covariates in formula string.
}
\keyword{datasets}

