% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_pdf}
\alias{create_jpeg}
\alias{create_pdf}
\alias{create_png}
\title{Create plots with confidence intervals of exposure variable}
\usage{
create_pdf(filename, models, rows = NULL, columns = NULL, labels = NULL)

create_jpeg(filename, models, rows = NULL, columns = NULL, ppi = 200)

create_png(filename, models, rows = NULL, columns = NULL, ppi = 200)
}
\arguments{
\item{filename}{Output filename}

\item{models}{List of models}

\item{rows}{One of "outcomes", "exposures", "adjustments"}

\item{columns}{One of "outcomes", "exposures", "adjustments"}

\item{labels}{Character vector with named elements (same as
\code{from_to} argument of \code{\link{translate}})}

\item{ppi}{Number of pixels per inch}
}
\value{
None.
}
\description{
Create plots with confidence intervals of exposure variable
}

