% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{distro}
\alias{distro}
\title{Summarize distribution of a continuous variable}
\usage{
distro(x, label = NULL, probs = NULL, digits = 2L)
}
\arguments{
\item{x}{Values of a continuous variable}

\item{label}{Label to use for continuous variable}

\item{probs}{Percentages for which to include percentiles}

\item{digits}{Number of decimal digits to use}
}
\value{
A data frame with columns "variable", "N", "NAs", "mean",
    "sd", "iqr", "min", percentiles, "max".  The "variable" column
    contains the name of the continuous variable.  The "N" columns
    contains the number of observations for the variable (including
    missing values).  The "NAs" column contains the number of
    missing values.  The "mean", "sd", "iqr", "min", and "max"
    columns contain the mean, standard deviation, inter-quartile
    range, minimum, and maximum of the variable, respectively.  The
    names of the percentile columns between the "min" and the "max"
    column all start with "p" followed by the percentage
    corresponding to the percentile.  The column containing the 5%
    percentile would be called "p05".  The column with the 50%
    percentile, or median, would be called "p50".
}
\description{
Summarize distribution of a continuous variable
}

