% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{distros}
\alias{distros}
\title{Summarize distribution of a continuous variable}
\usage{
distros(column_names, data, probs = NULL, digits = 2L)
}
\arguments{
\item{column_names}{Character vector with names of continuous variables}

\item{data}{Data frame containing variables `column_names`}

\item{probs}{Percentages for which to include percentiles}

\item{digits}{Number of decimal digits to use}
}
\value{
A data frame with columns "variable", "N", "NAs", "mean",
    "sd", "iqr", "min", percentiles, "max".  The "variable" column
    contains the `column_names`.  The "N" column contains the total
    number of observations per variable.  The "NAs" column contains
    the number of missing observations for each variable.  The
    "mean", "sd", "iqr", "min", and "max" variables contain the
    mean, standard deviation, inter-quartile range, minimum, and
    maximum of each variable, respectively.  The names of the
    percentile columns between the "min" and the "max" column all
    start with "p" followed by the percentage corresponding to the
    percentile.  The 5% percentile column would be called "p5",
    while the 50% percentile, or median, column column would be
    called "p50".
}
\description{
Summarize distribution of a continuous variable
}

