% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{escape_characters}
\alias{escape_characters}
\title{Escape characters from character class}
\usage{
escape_characters(x, character_class)
}
\arguments{
\item{x}{Character vector in which to escape characters}

\item{character_class}{Length-1 character vector defining a
"character class" (see \code{\link[base]{regex}})}
}
\value{
Character vector \code{x} in which the characters in
    \code{character_class} are escaped.
}
\description{
Escape characters from character class
}

