% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{filter_models}
\alias{filter_models}
\title{Extract models with selected outcomes, exposures, and adjustments.}
\usage{
filter_models(models, outcomes = NULL, exposures = NULL,
  adjustments = NULL, drop = FALSE, combine = NULL)
}
\arguments{
\item{models}{List of models}

\item{outcomes}{Character vector with names of outcome variables}

\item{exposures}{Character vector with names of exposure variables}

\item{adjustments}{List of character vectors with names of
adjustment variables}

\item{drop}{Drop matching models if TRUE (default FALSE)}

\item{combine}{If "and" models must match all of `outcomes`,
`exposures`, and `adjustments`.  If "or" (default) a single
match suffices.}
}
\value{
A list of models whose outcome, exposure, and adjustment
    match match `outcomes`, `exposures`, and `adjustments`.  The
    exact nature of the match depends on the values of `drop` and
    `combine`.
}
\description{
Extract models with selected outcomes, exposures, and adjustments.
}

