% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_combinations}
\alias{find_combinations}
\title{Find all combination of outcomes, exposures, and adjustments}
\usage{
find_combinations(outcomes = NULL, exposures = NULL, adjustments = NULL,
  by = NULL)
}
\arguments{
\item{outcomes}{Character vector of outcome variable names}

\item{exposures}{Character vector of exposures variable names}

\item{adjustments}{List of character vectors with adjustment
variable names}

\item{by}{Character vector defining how ties are broken}
}
\value{
List with sublists "outcomes", "exposures", "adjustments"
    of equal length.  Each triplet of corresponding elements of the
    three sublists forms a possible outcome-exposure-adjustment
    combination.
}
\description{
Find all combination of outcomes, exposures, and adjustments
}

