% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_layout}
\alias{find_layout}
\title{Find layout for plots}
\usage{
find_layout(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}
}
\value{
A list with elements "mat", "widths", "heights" that can be
    used as arguments to \code{\link[graphics]{layout}}.
}
\description{
Find layout for plots
}

