% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_layout_info}
\alias{find_layout_info}
\title{Convenience wrapper around \code{find_layout}}
\usage{
find_layout_info(models, rows, columns)
}
\arguments{
\item{models}{A list of models}

\item{rows}{One of "outcomes", "exposures", or "adjustments"}

\item{columns}{One of "outcomes", "exposures", or "adjustments"}
}
\value{
A list as returned by \code{\link{find_layout}}.
}
\description{
Convenience wrapper around \code{find_layout}
}

