% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_mapping_for_labels}
\alias{find_levels_for_labels}
\alias{find_mapping_for_labels}
\alias{find_variable_names_for_labels}
\title{Map labels as used in output of \code{summary} to variables/levels}
\usage{
find_mapping_for_labels(model)

find_variable_names_for_labels(labels, model)

find_levels_for_labels(labels, model)
}
\arguments{
\item{model}{Fitted model}
}
\value{
A data frame with columns "labels", "variables", and
    "levels".  The data frame represents a mapping from labels to
    variables and levels.  The data frame is used by
    \code{find_variable_names_for_labels} and
    \code{find_levels_for_labels} to map labels to variable names
    and levels respectively.  Labels that don't match any valid
    variable-level combination are be mapped to NA.  The variable
    name for the label "(Intercept)" is the label itself.  Labels
    of non-factor variables have an empty string as their "level".
}
\description{
Map labels as used in output of \code{summary} to variables/levels
}

