% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_margin_plots}
\alias{find_margin_plots}
\title{Find the numbers of plots at the margins}
\usage{
find_margin_plots(nrow, ncol, byrow = FALSE)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{byrow}{If FALSE (default) subplots are filled by columns,
otherwise by rows.}
}
\value{
A list with elements "bottom", "left", "top", and "right".
    Every element contains the numbers of the plots at the
    corresponding margin.
}
\description{
Find the numbers of plots at the margins
}

