% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{find_number_of}
\alias{find_number_of}
\title{Find number of outcomes, exposures, or adjustments}
\usage{
find_number_of(type, models)
}
\arguments{
\item{type}{One of "outcomes", "exposures", or "adjustments"}

\item{models}{List of models}
}
\value{
Number of variables of type \code{type}.
}
\description{
Find number of outcomes, exposures, or adjustments
}

