% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_page_row_column_values}
\alias{find_page_row_column_values}
\title{Match values of variables to different parts of a layout}
\usage{
find_page_row_column_values(variables, rows, columns)
}
\arguments{
\item{variables}{List with the same structure as the return value
of \code{\link{find_variables}}.}

\item{rows}{One of "outcomes", "exposures", "adjustments" but not
the same as `columns`}

\item{columns}{One of "outcomes", "exposures", "adjustments" but
not the same as `rows`}
}
\value{
The layout consists of pages, rows, and columns.  Every
    part of the layout represents one type of variable: outcomes,
    exposures, adjustments.  The function takes the values of the
    different types of variables (`variables`) as well as the
    correspondence between parts of the layout and the types of
    variables (`rows` and `columns`) and returns a list with
    elements "page", "row", "column" where every element contains
    the values of the type of variable that matches the
    corresponding part of the layout.
}
\description{
Match values of variables to different parts of a layout
}

