% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_pages_rows_columns}
\alias{find_pages_rows_columns}
\title{Map outcomes, exposures, and adjustments to pages, rows, and columns}
\usage{
find_pages_rows_columns(rows = NULL, columns = NULL)
}
\arguments{
\item{rows}{One of "outcomes", "exposures", or "adjustments"}

\item{columns}{One of "outcomes", "exposures", or "adjustments"}
}
\value{
List with elements "pages", "rows", and "columns".
}
\description{
Map outcomes, exposures, and adjustments to pages, rows, and columns
}
\details{
The \code{rows} and \code{columns} argument must either
    both be NULL or both be non-NULL.  If \code{rows} and
    \code{columns} are non-NULL, they must not have the same value.
}

