% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{find_position_in_layout}
\alias{find_position_in_layout}
\title{Find the position of a model in a layout}
\usage{
find_position_in_layout(model_number, layout_info)
}
\arguments{
\item{model_number}{Index of model in the list of models that is
used for plotting}

\item{layout_info}{Layout information as returned by
\code{\link{find_layout}}}
}
\value{
A list with elements "bottom", "left", "top", and "right"
    where an element is \code{TRUE} if the model is located in the
    respective part of the layout, otherwise the element is
    \code{FALSE}.
}
\description{
Find the position of a model in a layout
}

