% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{find_segment_limits}
\alias{find_segment_limits}
\title{Find x and y ranges of segments}
\usage{
find_segment_limits(segment)
}
\arguments{
\item{segment}{Data frame of segments as returned by
\code{\link{find_segments_to_plot}}}
}
\value{
A list with elements "x" and "y" giving the x and y ranges
    of the segments in \code{segments}.
}
\description{
Find x and y ranges of segments
}

