% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{find_segments_to_plot}
\alias{find_segments_to_plot}
\title{Find segments to plot (as confidence intervals)}
\usage{
find_segments_to_plot(model, type = "beta")
}
\arguments{
\item{model}{A fitted model object}

\item{type}{Confidence intervals to plot ("beta" or "OR")}
}
\value{
A data frame with columns "x0", "x1", "y0", "y1", and
    "midpoints" that can be used as arguments to
    \code{\link[graphics]{segments}}.  The midpoints correspond to
    position of the effect estimate within the confidence interval.
}
\description{
Find segments to plot (as confidence intervals)
}

