% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_variable_levels}
\alias{find_variable_levels}
\title{Find levels of variables}
\usage{
find_variable_levels(model, data)
}
\arguments{
\item{model}{Model object}

\item{data}{Dataset}
}
\value{
A named list containing the levels of variables in
    \code{model} as found in \code{data}.  Elements are named
    according to the corresponding variable.  Levels of non-factor
    variables are \code{NULL}.
}
\description{
Find levels of variables
}

