% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{formula.manyregs_model}
\alias{formula.manyregs_model}
\title{Convert model to formula.}
\usage{
\method{formula}{manyregs_model}(x, env = parent.frame())
}
\arguments{
\item{env}{Environment for formula}

\item{object}{Model to be converted to formula}
}
\value{
An object of class "formula" representing the model.
}
\description{
Convert model to formula.
}

