% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{freqs}
\alias{freqs}
\title{Tabulate categorical variables from data frame}
\usage{
freqs(column_names, data, digits = 2L)
}
\arguments{
\item{column_names}{Character vector with names of categorical variables}

\item{data}{Data frame containing categorical variables}

\item{digits}{Number of decimal digits to use for "\%" column}
}
\value{
A data frame with columns "variable", "category", "N", "n",
    "\%".  The "variable" column contains the column names of the
    categorical variables from `data`.  The "category" column
    contains the levels of the variables.  The "N" column contains
    the total number of observations per variable.  The "n" column
    contains the number of observations per category of each
    variable.  The "\%" column contains the percentage of
    observations that fall into a given category.
}
\description{
Tabulate categorical variables from data frame
}

