% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{is_categorical}
\alias{is_categorical}
\title{Is variable categorical in model?}
\usage{
is_categorical(variable, model)
}
\arguments{
\item{variable}{Name of covariate in \code{model}}

\item{model}{A model object}
}
\value{
Returns \code{TRUE} if \code{variable} is categorical in
    \code{model}, otherwise (if \code{variable} is continuous)
    returns \code{FALSE}.
}
\description{
Is variable categorical in model?
}

