% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{is_fitted}
\alias{is_fitted}
\title{Are models fitted?}
\usage{
is_fitted(models)
}
\arguments{
\item{models}{List of models}
}
\value{
TRUE for fitted models, FALSE for non-fitted models.
}
\description{
Are models fitted?
}

