% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{model_to_formula_string}
\alias{model_to_formula_string}
\title{Convert model to a formula-like string.}
\usage{
model_to_formula_string(x)
}
\arguments{
\item{x}{Model to be converted}
}
\value{
Formula-like character string representing the model.
}
\description{
Convert model to a formula-like string.
}

