% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{new_model}
\alias{new_model}
\title{Create a new model object of S3 class "manyregs_model".}
\usage{
new_model(outcome, exposure, adjustment, f, fname, template = NULL)
}
\arguments{
\item{outcome}{Name of outcome variable}

\item{exposure}{Name of exposure variable}

\item{adjustment}{Character vector with names of adjustment
variables}

\item{f}{Function for fitting the model}

\item{fname}{Name of fitting function}

\item{template}{Model to use as template}
}
\value{
Model object of S3 class "manyregs_model".
}
\description{
Create a new model object of S3 class "manyregs_model".
}

