% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.manyregs_model}
\alias{plot.manyregs_model}
\title{Plot model.}
\usage{
\method{plot}{manyregs_model}(x, ..., labels = NULL)
}
\arguments{
\item{x}{Model to be printed}

\item{labels}{Named character vector or named list.}

\item{\dots}{Ignored}
}
\value{
None.
}
\description{
Plot model.
}
\details{
If a variable in the model names one of the elements of
    \code{labels}, its name will be replaced by that element when
    used as a label.
}

