% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{print.manyregs_model}
\alias{print.manyregs_model}
\title{Print model.}
\usage{
\method{print}{manyregs_model}(x, ...)
}
\arguments{
\item{x}{Model to be printed}

\item{\dots}{Ignored}
}
\value{
Returns `x' invisibly.
}
\description{
Print model.
}

