% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{remove_slots}
\alias{remove_slots}
\title{Remove slots from a model.}
\usage{
remove_slots(model, slots)
}
\arguments{
\item{model}{Model from which to remove slots}

\item{slots}{Character vector of slot names}
}
\value{
Model from which named slots were removed.
}
\description{
Remove slots from a model.
}

