% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{sort_models}
\alias{sort_models}
\title{Sort a list of models}
\usage{
sort_models(models, by, outcomes = NULL, exposures = NULL,
  adjustments = NULL)
}
\arguments{
\item{models}{List of models}

\item{by}{Character vector defining how ties are broken}

\item{outcomes}{Character vector giving order of outcome variables}

\item{exposures}{Character vector giving order of exposure
variables}

\item{adjustments}{List of character vectors giving order of
adjustments}
}
\value{
A list of sorted models.
}
\description{
Sort a list of models
}
\details{
The \code{by} argument defines the order in which
    outcomes, exposures, and adjustments are use to sort the list
    of models.  For example, if \code{by} is \code{c("outcomes",
    "exposures", "adjustments")}, models are first sorted by
    outcomes.  Any ties are broken by sorting by exposures.  Any
    remaining ties are broken by sorting by adjustments.

    The sort order within outcomes (exposures, adjustments) is
    defined by the argument of the same name.  By default the sort
    order within outcomes (exposures, adjustments) is defined by
    the order of appearance of outcomes (exposures, adjustments) in
    the list of models.  By specifying only a subset of outcomes
    (exposures, adjustments) it is possible to obtain a sorted
    subset of models.
}

