% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{sort_models_for_plotting}
\alias{sort_models_for_plotting}
\title{Sort models for plotting in a layout of pages, rows, and columns}
\usage{
sort_models_for_plotting(models, rows = NULL, columns = NULL)
}
\arguments{
\item{models}{List of models}

\item{rows}{One of "outcomes", "exposures", or "adjustments"}

\item{columns}{One of "outcomes", "exposures", or "adjustments"}
}
\value{
A list of models sorted such that we can step through the
    list plotting one model after the other and every model will
    appear on the correct page in the correct row and column.
}
\description{
Sort models for plotting in a layout of pages, rows, and columns
}
\details{
The \code{rows} and \code{columns} arguments must not have
    the same value.
}

