% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{summarize_fitted_models}
\alias{summarize_fitted_models}
\title{Summarize a list of fitted models.}
\usage{
summarize_fitted_models(models)
}
\arguments{
\item{models}{List of models}
}
\value{
A single data frame summarizing the fitted models.
}
\description{
Summarize a list of fitted models.
}

