% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{summarize_models}
\alias{summarize_models}
\title{Summarize a list of models.}
\usage{
summarize_models(models)
}
\arguments{
\item{models}{Models to be summarized}
}
\value{
foobar
}
\description{
Summarize a list of models.
}

