% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{summarize_non_fitted_models}
\alias{summarize_non_fitted_models}
\title{Summarize a list of non-fitted models.}
\usage{
summarize_non_fitted_models(models)
}
\arguments{
\item{models}{List of models}
}
\value{
A data frame containing the printed representation of the
    models.
}
\description{
Summarize a list of non-fitted models.
}

