% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{summary.manyregs_model}
\alias{summary.manyregs_model}
\title{Summarize model.}
\usage{
\method{summary}{manyregs_model}(object, ...)
}
\arguments{
\item{object}{Model to be summarized}

\item{\dots}{Ignored}
}
\value{
A data frame summarizing the model.
}
\description{
Summarize model.
}

