% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{translate}
\alias{translate}
\title{Translate strings}
\usage{
translate(x, from_to)
}
\arguments{
\item{x}{Character vector}

\item{from_to}{Named character vector or named list}
}
\value{
A character vector of the same length as \code{x} with
    elements translated according to \code{from_to}.
}
\description{
Translate strings
}
\details{
If an element of \code{x} names one of the elements of
    \code{from_to}, it is replaced by that element.
}

