% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_estimates}
\alias{find_estimates}
\title{Extract estimates from fitted model}
\usage{
find_estimates(model)
}
\arguments{
\item{model}{Fitted model}
}
\value{
A data frame with columns "variable", "level", "beta",
    "se", "pvalue", "lcl", and "ucl" denoting the name of the
    variable, the corresponding level (in case of a factor), effect
    estimate, its standard error, the p-value, and the lower and
    upper 95\% confidence limits.

    Note that the confidence limits are based on the assumption of
    asymptotic normality of the effect estimates.  This assumption
    might be violated in small samples.
}
\description{
Extract estimates from fitted model
}

