% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_layout}
\alias{find_layout}
\title{Find layout for plots}
\usage{
find_layout(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}
}
\value{
A list with elements "mat", "widths", "heights" that can be
    used as arguments to \code{\link[graphics]{layout}}.  There are
    also elements "bottom", "left", "top", "right" that are integer
    vectors containing the numbers of the subplots at the bottom,
    left, top, and right margin, respectively.
}
\description{
Find layout for plots
}

