% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{find_plot_labels}
\alias{find_plot_labels}
\title{Find row, column, and page labels for model}
\usage{
find_plot_labels(model, rows, columns, labels = NULL)
}
\arguments{
\item{model}{A model object}

\item{rows}{One of "outcomes", "exposures", or "adjustments"}

\item{columns}{One of "outcomes", "exposures", or "adjustments"}

\item{labels}{Character vector with named elements (same as
\code{from_to} argument of \code{\link{translate}})}
}
\value{
A list with elements "row", "column", and "page" containing
    the row, column, and page labels for the model.
}
\description{
Find row, column, and page labels for model
}

