% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{find_variables}
\alias{find_variables}
\title{Find outcomes, exposures, adjustments from a list of models}
\usage{
find_variables(models)
}
\arguments{
\item{models}{List of model objects}
}
\value{
A list with elements "outcomes", "exposures", and
    "adjustments" representing the outcomes, exposures, and
    adjustments found among the models.
}
\description{
Find outcomes, exposures, adjustments from a list of models
}

