% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyregs.R
\name{fit_models}
\alias{fit_models}
\title{Fit models.}
\usage{
fit_models(models, data)
}
\arguments{
\item{models}{List of model objects}

\item{data}{Dataset to which models should be fitted}
}
\value{
List of fitted models.
}
\description{
Fit models.
}

