% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{manyregs-package}
\alias{manyregs}
\alias{manyregs-package}
\title{manyregs --- Flexible and concise regression analysis}
\description{
The manyregs package provides a concise and yet flexible way to
specify, fit, and summarize potentially large numbers of regression
models.
}

